feat: completely refactor remote GET / POST and local CRUD for BAHIS2 syncs
